Extract JavascriptBuilder into reusable class and abstract javascript base
